# api_docs.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Nigeria E-Voting API",
      "description": "A secure API for the Nigerian hybrid encryption-based voting system with USSD integration",
      "version": "1.0.0",
      "contact": {
        "name": "Nigeria Electoral Commission",
        "url": "https://evoting.gov.ng",
        "email": "support@evoting.gov.ng"
      },
      "license": {
        "name": "Proprietary",
        "url": "https://evoting.gov.ng/license"
      }
    },
    "servers": [
      {
        "url": "http://localhost:8080",
        "description": "Development Server"
      }
    ],
    "components": {
      "securitySchemes": {
        "BearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT",
          "description": "JWT token obtained from login endpoint"
        }
      },
      "schemas": {
...
        "UserCredentials": {
          "type": "object",
          "required": [
            "identifier",
            "password"
          ],
          "properties": {
            "identifier": {
              "type": "string",
              "description": "NIN, VIN, or phone number",
              "example": "12345678901"
            },
            "password": {
              "type": "string",
              "format": "password",
              "example": "securepassword123"
            }
          }
        },
...
        "VoterRegistration": {
          "type": "object",
          "required": [
            "nin",
            "vin",
            "phoneNumber",
            "dateOfBirth",
            "password",
            "fullName",
            "pollingUnitCode",
            "state",
            "gender",
            "lga",
            "ward"
          ],
          "properties": {
            "nin": {
              "type": "string",
              "description": "11-character National Identification Number",
              "example": "12345678901"
            },
            "vin": {
              "type": "string",
              "description": "19-character Voter Identification Number",
              "example": "1234567890123456789"
            },
            "phoneNumber": {
              "type": "string",
              "description": "Phone number for MFA",
              "example": "+2348012345678"
            },
            "dateOfBirth": {
              "type": "string",
              "format": "date",
              "description": "Date of birth for verification",
              "example": "1990-01-01"
            },
            "password": {
              "type": "string",
              "format": "password",
              "description": "User password",
              "example": "securepassword123"
            },
            "fullName": {
              "type": "string",
              "description": "Full name of the voter",
              "example": "John Doe"
            },
            "pollingUnitCode": {
              "type": "string",
              "description": "Assigned polling unit code",
              "example": "PU001"
            },
            "state": {
              "type": "string",
              "description": "State of residence",
              "example": "Lagos"
            },
            "gender": {
              "type": "string",
              "enum": [
                "male",
                "female"
              ],
              "description": "Gender of the voter",
              "example": "male"
            },
            "lga": {
              "type": "string",
              "description": "Local Government Area",
              "example": "Ikeja"
            },
            "ward": {
              "type": "string",
              "description": "Ward within the LGA",
              "example": "Ward 1"
            }
          }
        },
...
        "ForgotPassword": {
          "type": "object",
          "required": [
            "email"
          ],
          "properties": {
            "email": {
              "type": "string",
              "format": "email",
              "description": "Registered email for password reset",
              "example": "user@example.com"
            }
          }
        },
        "ResetPassword": {
          "type": "object",
          "required": [
            "token",
            "newPassword"
          ],
          "properties": {
            "token": {
              "type": "string",
              "description": "Password reset token",
              "example": "a1b2c3d4e5f67890abcdef1234567890"
            },
            "newPassword": {
              "type": "string",
              "format": "password",
              "description": "New password",
              "example": "newSecurePass123"
            }
          }
        },
        "LoginResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": true
            },
            "message": {
              "type": "string",
              "example": "Login successful"
            },
            "data": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string",
                  "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjEyMzQ1Njc4OTAiLCJyb2xlIjoidm90ZXIiLCJpYXQiOjE2NTA5ODExMjksImV4cCI6MTY1MDk4NDcyOX0.DVI9S5pZ5EWTnO3vCsfl4xOFLVeYs"
                },
                "voter": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "format": "uuid",
                      "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                    },
                    "nin": {
                      "type": "string",
                      "example": "12345678901"
                    },
                    "vin": {
                      "type": "string",
                      "example": "1234567890123456789"
                    },
                    "phoneNumber": {
                      "type": "string",
                      "example": "+2348012345678"
                    }
                  }
                },
                "requiresMfa": {
                  "type": "boolean",
                  "example": false
                }
              }
            }
          }
        },
        "ErrorResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "example": false
            },
            "code": {
              "type": "string",
              "example": "VALIDATION_ERROR",
              "description": "Error code for programmatic handling"
            },
            "message": {
              "type": "string",
              "example": "Invalid credentials"
            },
            "error": {
              "type": "string",
              "description": "Error message",
              "example": "Invalid input data"
            },
            "details": {
              "type": "object",
              "description": "Additional error details",
              "example": {
                "field": "email",
                "message": "Invalid email format"
              }
            },
            "timestamp": {
              "type": "string",
              "format": "date-time",
              "description": "Error timestamp",
              "example": "2023-02-25T10:30:00Z"
            }
          }
        },
...
      },
...
    },
    "security": [
      {
        "BearerAuth": []
      }
    ],
...
    "paths": {
...
      "/api/v1/auth/register": {
        "post": {
          "summary": "Register a new voter",
          "tags": [
            "Authentication"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "nin",
                    "vin",
                    "phoneNumber",
                    "dateOfBirth",
                    "password",
                    "fullName",
                    "pollingUnitCode",
                    "state",
                    "gender",
                    "lga",
                    "ward"
                  ],
                  "properties": {
                    "nin": {
                      "type": "string",
                      "description": "11-character National Identification Number",
                      "example": "12345678901"
                    },
                    "vin": {
                      "type": "string",
                      "description": "19-character Voter Identification Number",
                      "example": "1234567890123456789"
                    },
                    "phoneNumber": {
                      "type": "string",
                      "description": "Phone number for MFA",
                      "example": "+2348012345678"
                    },
                    "dateOfBirth": {
                      "type": "string",
                      "format": "date",
                      "description": "Date of birth for verification",
                      "example": "1990-01-01"
                    },
                    "password": {
                      "type": "string",
                      "format": "password",
                      "description": "User password"
                    },
                    "fullName": {
                      "type": "string",
                      "description": "Full name of the voter",
                      "example": "John Doe"
                    },
                    "pollingUnitCode": {
                      "type": "string",
                      "description": "Assigned polling unit code",
                      "example": "PU001"
                    },
                    "state": {
                      "type": "string",
                      "description": "State of residence",
                      "example": "Lagos"
                    },
                    "gender": {
                      "type": "string",
                      "enum": [
                        "male",
                        "female"
                      ],
                      "description": "Gender of the voter",
                      "example": "male"
                    },
                    "lga": {
                      "type": "string",
                      "description": "Local Government Area",
                      "example": "Ikeja"
                    },
                    "ward": {
                      "type": "string",
                      "description": "Ward within the LGA",
                      "example": "Ward 1"
                    }
                  },
                  "$ref": "#/components/schemas/VoterRegistration"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Voter registered successfully",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "Voter registered successfully"
                      },
                      "data": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid",
                            "example": "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
                          },
                          "nin": {
                            "type": "string",
                            "example": "12345678901"
                          },
                          "vin": {
                            "type": "string",
                            "example": "1234567890123456789"
                          },
                          "phoneNumber": {
                            "type": "string",
                            "example": "+2348012345678"
                          }
                        }
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input parameters",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "409": {
              "description": "Voter already exists",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/auth/login": {
        "post": {
          "summary": "Login a voter",
          "tags": [
            "Authentication"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "identifier",
                    "password"
                  ],
                  "properties": {
                    "identifier": {
                      "type": "string",
                      "description": "NIN, VIN, or phone number",
                      "example": "12345678901"
                    },
                    "password": {
                      "type": "string",
                      "format": "password"
                    }
                  },
                  "$ref": "#/components/schemas/UserCredentials"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Login successful",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/LoginResponse"
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input parameters",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "401": {
              "description": "Invalid credentials",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "403": {
              "description": "Account inactive",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
...
      "/api/v1/auth/forgot-password": {
        "post": {
          "summary": "Request password reset",
          "tags": [
            "Authentication"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "email"
                  ],
                  "properties": {
                    "email": {
                      "type": "string",
                      "format": "email"
                    }
                  },
                  "$ref": "#/components/schemas/ForgotPassword"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Password reset email sent",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "Password reset instructions sent to your email"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input parameters",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "404": {
              "description": "Email not found",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
      "/api/v1/auth/reset-password": {
        "post": {
          "summary": "Reset password with token",
          "tags": [
            "Authentication"
          ],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "token",
                    "newPassword"
                  ],
                  "properties": {
                    "token": {
                      "type": "string"
                    },
                    "newPassword": {
                      "type": "string",
                      "format": "password"
                    }
                  },
                  "$ref": "#/components/schemas/ResetPassword"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "Password reset successful",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "success": {
                        "type": "boolean",
                        "example": true
                      },
                      "message": {
                        "type": "string",
                        "example": "Password reset successful"
                      }
                    }
                  }
                }
              }
            },
            "400": {
              "description": "Invalid input parameters",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "401": {
              "description": "Invalid or expired token",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            },
            "500": {
              "description": "Server error",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/ErrorResponse"
                  }
                }
              }
            }
          }
        }
      },
...
    }
  }